requirements.txt
asgiref==3.8.1
boto3==1.38.41
botocore==1.38.41
certifi==2025.6.15
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
Django==5.2.3
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-extensions==4.1
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
idna==3.10
jmespath==1.0.1
msgpack==1.1.1
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
redis==6.2.0
requests==2.32.4
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
